hdl/top/afc_v3/*: add new top files with 2_to_1 acquisiton core
This top file is the same as the afc_v3/dbe_bpm_dsp_fmc130m_4ch
with the difference that we use the new wb_acq_core_2_to_1_mux
for simultaneous acquisition of BPM1 and BPM2.

This was suppose to fix github issue #31, but there is a
problem when reading from the PCIe DDR3 memory space
whle still acquiring samples. The PCIe core hangs
when asked to read/write to/from DDR3 and outputs
FF's forever.