wb_acq_core/*: add trigger support to acquisition module
Now, we can select a variable number of pre-trigger samples
and a variable number of post-trigger samples to acquire.
We employ a basic trigger acquisition in which the samples
are continuously written to DDR memory and, on detecting a
trigger, tag this DDR address event, and write the number
of remaining samples (post-samples)to DDR memory. The
acquisition readout logic, such as DDR memory sections
for multiple acquisition rates and DDR address wrapping,
is done by software.