wb_acq_core/*: retag data_id DPRAM samples to PRE_TRIG
We use the data ID tag to count the valid samples
written do DDR. As the DPRAM only outputs valid
samples, we will almost certainly have WAIT_TRIG
and POST_TRIG samples only, and not PRE_TRIG samples.
This means that downstream modules will not count these
samples as valid.

The fix was to retag the samples from WAIT_TRIG to
PRE_TRIG.