wb_acq_core/*: fix wrong acq_chan_ctl reg synch to fs_clk_i
We were using a synching this register with ext_clk_i,
which lead to timing errors. We are already synching
this reg to ext_clk inside wb_acq_core. So, indeed,
the correct clock to use is fs_clk_i, as every module,
except a few ones are clocked to ext_clk_i.