wb_acq_core/acq_fc_fifo.vhd: fix data_id channel selection
For channels that occupy a number of FIFOs different
than the maximum (e.g., for number of FIFOs equal to
4 with data width of 64 bits, a channel having data
width equal to 128 would occupy only 2 FIFOs), the data_id
selection was wrong.

For a channel occupying 2 FIFOs, the data_id index
would always select 4, which is correct for the channel
occupying all FIFOs, but wrong for the channel with 2 FIFOs.