testbench/*/wb_acq_core_test/**: delay DDR3 reading so RAW hazards do not occur
This happened as the AXI data mover has a
significant delay between the AXIS and AXI interface.
Because of this, the last words were not written to
DDR3 memory as the reading started.

This is solved adding a fixed delay between the writing
and reading parts. On a typical application, the CPU
would add much more latency.