modules/*/wb_acq_core/*: add AXIS support for multishot and trigger cases
Now, we use the datamover in indeterminate BTT mode,
in which we can send only one packet to transfer up
to 2^23-1 bytes. If we need to abort it earlier, we can
by asserting the TLAST signal ahead of time.

For the trigger cases, in which, we must write new data
for an indeterminate period of time (until the trigger
actually comes), we just detect the end of the 2^23-1 bytes
and start another one. When the trigger indeed happens,
we send a TLAST signal on the data stream buffer and the
current transaction is finished.