modules/*/wb_acq_core/*: shift samples as soon as possible to avoid race conditions
We need to use the lmt_curr_chan_id_i signal to address
our current channel array, as we need to possible modify
the samples before sending it to internal modules.

We can't wait for lmt_valid_i signal, as the internal
modules use it as well to acquire the number of samples.

Anyway, this operation is safe, as only after the lmt_valid_i
signal propagates to the other modules is that we will actually
use it.