modules/pcie/*: fix DDR depth generic by importing from ipcores_pkg
The previous DDR depth was hardcoded as 30,
which is right for kc705 board, but not for
AFC.

This fixes that by using the ipcores_pkg
definitions which are respective to the
DDR definitions of the used board.

This should help github issue #45,
but in my tests it didn't solve it
just yet. Nonethless, this was wrong
anyway.