modules/*/*trigger_iface: fix bogus receiving pulse event when transmitting
IOBUF primitives always have the "input"
signal available for further use. This means
that even when transmitting signals (buffer
is output) the input signal will have a copy
of the transmitted pulse.

This is principle is not a problem, but we
were using this signal to count received events,
so we changed this to use this signal only if
the buffer direction is set to input.