modules/*/wb_trigger_iface: fix trigger logic to only drive when sending data
In order to use the wired-OR logic of the
MLVDS standard, we need to output data only
when sending data.

If we are in output state we drive the data
pin to HI and use the direction pin as data,
so we can release the bus when not sending.

If we are in input state, we just use the data/
direction pins as usual.