modules/*/wb_acq_core/*: fix wrong address wrapping
Previously we were comparing the end address
with '=', but in most cases the ddr end address
is not multiple of the internal address. So, the
comparison would never trigger and never wrapping
the counter to the initial address, but to 0.

This caused long running acquisitions (e.g., external
trigger) to overwrite results of other channels.