modules/*/wb_acq_core/*: fix possible AXI mem region overflow
Previously, we were always setting the BTT as
2^22. If we needed to finish early, we just ended
the transaction with TLAST.

This posed a problem for long-running acquisition
(triggered), in that the AXI datamover would
continue to write even after the memory region,
because the BTT was set to 2^22.

Now, we always set it to the maximum of the memory
region.