dbe_bpm_gen.vhd: Include trigger to sync RFFE switching clock phase
The recv trigger channel 19 is connected now to the core that generates the RFFE switching clock (swap_freqgen) and is used to synchronize its phase.
The switching counter is reset upon a high-state detection on the trigger line.
One can use a MLVDS external trigger to sync all boards at the same time