top/*/afcv3/*: add snippet to merge all .xdc files in correct order
This is needed as we are not relying on XCI workflow
for PCIe and DDR cores anymore. So, the .xdc processing
order for ddr_core.xdc and pcie_core.xdc are the same
as the dbe_bpmX.xdc one. This might cause constraints
to not being created before using them.