modules/position_calc: fix width calculation to always fit size
Previously, if the size was a power of 2, an
incorrect width was calculated and depending
on the logic, errors would occur.

For instance, in decimation strobe
calculation, if the size is a power of 2,
say 32, the width would be calculated as
5. But with this width, we can't represent
unsigned 32.

This fixed the bug when decimation rates or
decimation strobes are power of 2.