modules/wb_bpm_swap: fix freqgen synch to only zero at en_i = '1'
When synch'ing with an external pulse, we need to
zero the counter only at the en_i signal. Otherwise,
the signal frequency would be enlarged.

With this fix this wouldn't happen, but we now have
a jitter of up to the en_i rate. In our case, this is
either 0, 4, or 8 ADC clock samples (~4ns).